Apply Autofilter vba code
The Apply Autofilter VBA function automates the application of Auto filter to a specified range in an Excel worksheet using VBA. It enables users to filter data based on defined criteria, simplifying data analysis and management. This function allows for dynamic sorting and display of relevant information by specifying the range and filter conditions, streamlining the process of working with large datasets.