Create New Workbook vba code
The CreateNewWorkbookVBA function is a VBA tool used to automate the creation of a new Excel workbook. It generates a blank workbook, which can be used for starting new projects, generating reports, or creating templates. This function simplifies the process of creating new workbooks, making it efficient to manage multiple workbooks within Excel.